+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 6.3.0
+## Added
+- setting Run document macros
+
+## Changed
+- compatible with Confluence 9.5.2
+- updated list supported formats, see [ONLYOFFICE/document-formats v3.0.0](https://github.com/ONLYOFFICE/document-formats/releases/tag/v3.0.0)
+- fix CVE-2025-27820
+
 ## 6.2.0
 ## Changed
 - com.onlyoffice:documentserver-sdk-java:1.4.0